+++ lib/test_rail.dart
@@ -1,15 +1,15 @@
 import 'package:http/http.dart' as http;
 import 'flutter_testrail.dart';
 
-class TestRail {
-  TestRail._({
+class FlutterTestRail {
+  FlutterTestRail._({
     required this.client,
   });
-  final TestRailHttpClient client;
+  final FlutterTestRailHttpClient client;
 
-  static late TestRail instance;
+  static late FlutterTestRail instance;
 
-  static TestRail configure({
+  static FlutterTestRail initialize({
     required String password,
 
     /// The url that points to the TestRail server => https://example.testrail.com
@@ -17,8 +17,8 @@ class TestRail {
     required String username,
     http.Client? client,
   }) {
-    return instance = TestRail._(
-      client: TestRailHttpClient(
+    return instance = FlutterTestRail._(
+      client: FlutterTestRailHttpClient(
         client: client,
         password: password,
         serverDomain: serverDomain,