+++ lib/test_section.dart
@@ -30,7 +30,7 @@ class TestSection {
       );
 
   static Future<TestSection> get(int sectionId) async {
-    final response = await TestRail.instance.client
+    final response = await FlutterTestRail.instance.client
         .request('/get_section/$sectionId', RequestMethod.get);
     return TestSection.fromJson(response!);
   }