+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.0
+
+********** BREAKING CHANGES *********
+
+1 . TestRail.configure() is now FlutterTestRail.initialize()
+
+2. Add New Method to add Single test Result by : TestResult.addTestResult(9,statusId:1)
+
+3. Now you can also add multiple Test results in one time : 
+  ### TestResult.addRunResults(3,addTestResults:testStatusResult);
+
 ## 0.0.4
 Add Support for latest dart sdk: '>=3.0.1 <4.0.0'
 ## 0.0.3